WARNING: This version of Cobblemon Intuition is NOT COMPATIBLE with Cobblemon 1.6.1. Please use version 0.5.0 of this mod if you are playing on 1.6.1.
+ Upgraded project resources
- Removed unused references
- Patched a bait-related exploit with Cobblemon fishing
+ Now compatible with Cobblemon 1.7

Known bug: When Intuition expires, the Pokemon that is fished up is not the one that was previously predicted by Intuition.

This is caused by Intuition only modifying the base Cobblemon fishing code whenever the Sonar (Intuition) potion effect is applied to the player. Thus, the fishing result is no longer overwritten at reel-in.